refactor: create utility functions file
